added PR changes
