+++ ChatSX/components/ParseDeepLink/models/DeepLinkRequest.ts
@@ -1,0 +1,6 @@
+export interface DeepLinkRequest {
+  uri: string;
+  userName: string;
+  uuid: string;
+  isGuest: boolean;
+}
