+++ ChatSX/components/ParseDeepLink/models/DeepLinkResponse.ts
@@ -1,10 +1,3 @@
-export interface DeepLinkRequest {
-  uri: string;
-  userName: string;
-  uuid: string;
-  isGuest: boolean;
-}
-
 export interface DeepLinkResponse {
   success: boolean;
   errorMessage?: string;
