+++ ChatSX/components/ParseDeepLink/models/index.tsx
@@ -1,0 +1,4 @@
+import {DeepLinkRequest} from './DeepLinkRequest';
+import {DeepLinkResponse} from './DeepLinkResponse';
+
+export {DeepLinkRequest, DeepLinkResponse};
