+++ ChatSX/components/Poll/models/CreatePoll.ts
@@ -1,0 +1,4 @@
+export interface CreatePoll {
+  navigation: any;
+  route: any;
+}
