+++ ChatSX/components/Poll/models/CreatePollProps.ts
@@ -1,0 +1,5 @@
+import {CreatePollCallbacks} from './CreatePollCallbacks';
+import {CreatePollStateProps} from './CreatePollStateProps';
+
+export declare type CreatePollProps = CreatePollStateProps &
+  CreatePollCallbacks;
