+++ ChatSX/components/Poll/models/Poll.ts
@@ -1,0 +1,6 @@
+export interface Poll {
+  id: string;
+  isSelected: boolean;
+  percentage: number;
+  noVotes: number;
+}
