+++ ChatSX/components/Poll/models/PollConversationUIProps.ts
@@ -1,0 +1,5 @@
+import {PollConversationViewCallbacks} from './PollConversationViewCallbacks';
+import {PollConversationViewState} from './PollConversationViewState';
+
+export declare type PollConversationUIProps = PollConversationViewState &
+  PollConversationViewCallbacks;
