+++ ChatSX/components/Poll/models/PollConversationViewCallbacks.ts
@@ -1,0 +1,13 @@
+export interface PollConversationViewCallbacks {
+  onNavigate: (val: string) => void;
+  setSelectedPollOptions: (pollIndex: any) => void;
+  addPollOption: () => void;
+  submitPoll: () => void;
+  setShowSelected: (show: boolean) => void;
+  setIsAddPollOptionModalVisible: (visible: boolean) => void;
+  setAddOptionInputField: (inputField: string) => void;
+  openKeyboard: () => void;
+  longPressOpenKeyboard: () => void;
+  stringManipulation: () => string;
+  resetShowResult: () => void;
+}
