+++ ChatSX/components/Poll/models/PollConversationViewProps.ts
@@ -1,0 +1,7 @@
+export interface PollConversationViewProps {
+  navigation: any;
+  item: any;
+  isIncluded: boolean;
+  openKeyboard: () => void;
+  longPressOpenKeyboard: () => void;
+}
