+++ ChatSX/enums/ChatroomActions.ts
@@ -1,0 +1,12 @@
+export enum ChatroomActions {
+  VIEW_PARTICIPANTS = 2,
+  SHARE = 3,
+  JOIN_CHATROOM = 4,
+  MUTE_NOTIFICATIONS = 6,
+  UNMUTE_NOTIFICATIONS = 8,
+  LEAVE_CHATROOM = 9,
+  LEAVE_SECRET_CHATROOM = 15,
+  VIEW_PROFILE = 21,
+  BLOCK_MEMBER = 27,
+  UNBLOCK_MEMBER = 28,
+}
