+++ ChatSX/shareUtils/models/ShareChatroomRequest.ts
