+++ ChatSX/shareUtils/models/index.tsx
@@ -1,0 +1,3 @@
+import {ShareChatroomRequest} from  './ShareChatroomRequest';
+
+export {ShareChatroomRequest};
